requirements.txt
Django~=2.2.4
Pillow~=6.2.0
gunicorn~=19.9.0
WhiteNoise~=4.1.3
Keras~=2.2.5
numpy~=1.16.4
matplotlib~=3.1.0
opencv-python==4.1.0.25
imageio~=2.5.0
gunicorn~=19.9.0
cloudinary
torch
torchvision